Flask_Blog/requirements.txt
bcrypt==3.1.7
cffi==1.12.3
Click==7.0
cycler==0.10.0
et-xmlfile==1.0.1
Flask==1.1.1
Flask-Admin==1.5.3
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
flask-mongoengine==0.9.5
Flask-PyMongo==2.3.0
Flask-WTF==0.14.2
flaskwebgui==0.0.7
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.10.1
joblib==0.14.0
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==2.1.0
mongoengine==0.18.2
numpy==1.16.3
openpyxl==2.6.0
pandas==0.23.4
psutil==5.6.3
pycparser==2.19
pymongo==3.9.0
pyparsing==2.4.5
python-dateutil==2.8.1
pytz==2019.3
scikit-image==0.15.0
scikit-learn==0.19.1
scipy==1.2.0
six==1.12.0
sklearn-crfsuite==0.3.6
Werkzeug==0.16.0
whichcraft==0.6.1
WTForms==2.2.1
xlrd==1.2.0
